Added decimal support to asset model contracts

+++ marketplace/backend/dapp/mercata-base-contracts/Templates/Assets/UTXO.sol
@@ -29,7 +29,7 @@ abstract contract UTXO is Asset {
     }
 
     // Quantity is already checked by transferOwnership function
-    function _transfer(address _newOwner, uint _quantity, bool _isUserTransfer, uint _transferNumber, uint _price) internal override {
+    function _transfer(address _newOwner, uint _quantity, bool _isUserTransfer, uint _transferNumber, decimal _price) internal override {
         require(status != AssetStatus.PENDING_REDEMPTION, "Asset is not in ACTIVE state.");
         require(status != AssetStatus.RETIRED, "Asset is not in ACTIVE state.");
         require(checkCondition(), "Condition is not met");
